Intermediate commit

+++ src/FoodOrderSystem.Domain/Commands/RemoveRestaurant/RemoveRestaurantCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using FoodOrderSystem.Domain.Model;
+using FoodOrderSystem.Domain.Model.Dish;
+using FoodOrderSystem.Domain.Model.DishCategory;
 using FoodOrderSystem.Domain.Model.Restaurant;
 using FoodOrderSystem.Domain.Model.User;
 using System;
@@ -10,10 +12,14 @@ namespace FoodOrderSystem.Domain.Commands.RemoveRestaurant
     public class RemoveRestaurantCommandHandler : ICommandHandler<RemoveRestaurantCommand, bool>
     {
         private readonly IRestaurantRepository restaurantRepository;
+        private readonly IDishCategoryRepository dishCategoryRepository;
+        private readonly IDishRepository dishRepository;
 
-        public RemoveRestaurantCommandHandler(IRestaurantRepository restaurantRepository)
+        public RemoveRestaurantCommandHandler(IRestaurantRepository restaurantRepository, IDishCategoryRepository dishCategoryRepository, IDishRepository dishRepository)
         {
             this.restaurantRepository = restaurantRepository;
+            this.dishCategoryRepository = dishCategoryRepository;
+            this.dishRepository = dishRepository;
         }
 
         public async Task<Result<bool>> HandleAsync(RemoveRestaurantCommand command, User currentUser, CancellationToken cancellationToken = default)
@@ -26,6 +32,14 @@ namespace FoodOrderSystem.Domain.Commands.RemoveRestaurant
 
             if (currentUser.Role < Role.SystemAdmin)
                 return FailureResult<bool>.Forbidden();
+
+            var dishCategories = await dishCategoryRepository.FindByRestaurantIdAsync(command.RestaurantId, cancellationToken);
+            if (dishCategories != null && dishCategories.Count > 0)
+                return FailureResult<bool>.Create(FailureResultCode.RestaurantContainsDishCategories);
+
+            var dishes = await dishRepository.FindByRestaurantIdAsync(command.RestaurantId, cancellationToken);
+            if (dishes != null && dishes.Count > 0)
+                return FailureResult<bool>.Create(FailureResultCode.RestaurantContainsDishes);
 
             await restaurantRepository.RemoveAsync(command.RestaurantId, cancellationToken);
